Initial import of GcodeAnalyzer project.
The objective of this project is calculate statistics about the Gcode, such that the performance of executing that Gcode on a particular machine may be evaluated before the print.  For example, if we know that the Ultimaker 2 can only safely extrude PLA at rates of 8 mm^3/S, then we can use this analyzer to check the rate of extrusion expressed by the Gcode, and compare it to our threshold of 8.  If the Gcode expresses a number that's too high, the slicer parameters should be tuned to bring that particular stat in compliance with the printer's capability.