Partially implemented object storage UI. Object data is stored in [ProjectRoot]/Data/ObjectStorage. each folder in this directory represents a different object. In each of those folders there is an Info.xml which contains some basic info about the object, such as its name, gui, author, and description. There are some additional sub folders in these folders as well:
Media: Downloaded images and other promo data.
GCode: This is where downloaded gcode for this object is stored.
Extra: Extra data that the author wants to make available to the user, such as source model file.

The folders contents is downloaded and stored for quick access.

The Info.xml can be loaded and saved from the ObjectData class, which is a code representation of the xml data.