Initial attempt at calculating payment dates.
As an aside, while implementing this, I discovered that our disbursement crontab isn't doing what I thought.  It was intended to execute every three days, but that doesn't hold true across month boundaries.  Crontab's day field is defined as "day of month" and therefore, if we always divide the "day of month" by three, then the count gets reset for every month.

I'm going to have to implement a script that executes every day at the desired time, and the script itself will have to figure out whether it's on the interval day (every three).  I'm basically doing that in this NextPaymetDate() calculation, but taking the number of days since epoch, and modulo'ing by 3.